Travis configuration for Mac/Windows clients build. (#208)
Added Travis jobs that generate a Mac package and a Windows installer for Fleetspeak. Both installers are "generic" installers, assuming that the Fleetspeak configuration (a file for Mac, registry entries for Windows) will be written externally and available at the time of the installation.

Along the way: using go 1.13 on Travis, along with go modules (pinned down versions to make builds reproducible).