Added DeletePendingMessages admin RPC call. (#215)
* Added DeletePendingMessages database call and implemented it for MySQL and sqlite.
* Added a corresponding admin RPC call.
* Added a corresponding method to the Python client.

This should help in cases when a client has a number of messages in the queue that can potentially take a long time to be processed.

At the moment there's no way and no requirement to specify the service to delete the pending mesages for. This should be fine for the first iteration.