Fix permission error with wnix sockets.
In the fix for https://github.com/golang/go/issues/35033 Go started creating
read-only files on Windows when no read permission bit was requested in the
OpenFile call. This was not documented in the release notes (but see
https://go-review.googlesource.com/c/go/+/223957/).

Deal with the immediate issue by adding owner-read and owner-write permissions
in our WriteFile which creates the file.

Also add a workaround for the fact that previous version of Fleetspeak have left
behind read-only files - simply clear the read-only attribute if the file
already exists.