Fix missing mode bits in ioutil.WriteFile
I was previously under the impression that the third param has no effect when
the file already exists but empirically this is not true on Windows - the
read-only bit is modified.

This wouldn't have broken anything because we just clear the bit before writing
anyway, but let's not leave a read-only file lying around.