Update in response to wnixsocket review comments (#236)
* Remove mode param from wnixsocket.Listen

... and instead hard-code the permissions which the wnixsocket implementation
requires. This is the only value for the permission bits which makes sense.

* Remove custom Windows Chmod code

The referenced PR has now been merged in go-acl. In the meantime this has also
been merged:

https://github.com/hectane/go-acl/commit/0a8ab9f24f8b1e9ab338540d28520a22b132828d#diff-b0c0f5d28453df828e70bcd9c0b16597

Which provides a convenience for administrators.

* Fix missing mode bits in ioutil.WriteFile

I was previously under the impression that the third param has no effect when
the file already exists but empirically this is not true on Windows - the
read-only bit is modified.

This wouldn't have broken anything because we just clear the bit before writing
anyway, but let's not leave a read-only file lying around.

* Improve some comments in wnixsocket_windows.go

- Fix reference to the "above WriteFile" which is actually below.

- Add comments on the third param to ioutil.WriteFile, since that has caused
  some confusion

- Be more specific on the function of the Chmod call (we are ensuring the ACLs
  are correct here - "mode" sounds like it includes the read-only attribute, but
  that is now handled by WriteFile).