Removing default port for collector and flag for choosing Collector
Instead of introducing flag to allow choosing a stats collector, only a bind address is chosen (host and port). Once an address is chosen in components.textproto under stats_config, the PrometheusStatsCollector struct is initialized and metrics are exported to <host>:<port>/metrics. If a port is not chosen, then no metrics are gathered (effectively, a noopStatsCollector is used).
Using a bind address instead of a port alone was suggested in comments
by @bgalehouse.