Use tempdir to detect our own user and group SID for socket checks
On WINE I find that this ownership check fails even when the socket's parent
directory is created in the same process where the check is done. We are also
seeing the error on our Windows fleet - we cannot prove that the directory is
created in the same process on the real machines, but it is certainly created by
Fleetspeak, so it's very likely the same issue.

It looks like the only reason we were using the group in the first place was for
the reason described in the comment - this commit just implements the
alternative workaround that was described there. Changing to use this workaround
allows us to check the owner SID as well as the group SID.

This change made the issue go away on WINE, maybe it will on the fleet too.