Changes in fetchResourceUsageRecordsTest
The tests now doesn't depend on a fake timestamp that we provide it, but
rather we check the relative timestamps. We do it by saving the
timestamps before and after we record the data into the datastore, and
then we use that as the time range for the fetch operation.