Don't quit process, return from svc Execute function
Per https://godoc.org/golang.org/x/sys/windows/svc#Handler we are supposed to
return from the Execute function. It is suspected that not doing so leads to the
service manager reporting "The FleetspeakService service terminated
unexpectedly"

This implements a graceful shutdown, and triggers it in between sending the
StopPending message and returning from the Execute callback. No such trigger for
the graceful shutdown is implemented on Unix.