Make the https component of the fleetspeak server optional.
This change makes it possible to run a pure fleetspeak admin server (without
running a HTTPS server as well).

The purpose of this change is to enable a distributed open source setup,
consisting of multiple HTTPS servers behind a load balancer and multiple admin
servers behind a second load balancer.

The pure admin server needs to be able to use a http based notifier, without
running a http based notification listener as well. For this, a new config
option is introduced.