Refactored intprocess.KillProcess() into KillProcessByPid(). (#342)
This fixes #341

The issue was that building os.Process object as &os.Process{pid: 42}, even though it worked in previous Go versions, effectively breaks encapsulation and is not going to work with Go 1.17 at least on Windows. The correct solution is to use os.FindProcess.

Newly introduced function - KillProcessByPid() - should be used in cases where only a process pid is available. Whenever a fully-featured os.Process object is available, its Kill() method is preferred.