execution: Pass context to `busySleep()`, pull context further up
The newly created context is going to be Done() when `e.dead` is closed.

It's better to use a context here.  This way, the cancelation condition
is abstract to `busySleep()`, and it's more obvious in the calling location
what the actual cancelation condition is.  It removes the boolean protocol
between caller and `busySleep()`.

PiperOrigin-RevId: 601057283