Bind both statsRoutine and heartbeatMonitorRoutine to the process lifetime
Both functions should exit as soon as the process has exited.

statsRoutine is reading resource usage statistics from /proc/$PID,
and can not possibly work if the PID is already gone, so an earlier
limit is appropriate.

heartbeatMonitorRoutine is watching heartbeats in an indirect way,
but its job is to kill the process if it does not heartbeat correctly.
That is also not possible without the PID still existing.

So both routines should be bound to the process lifetime and nothing longer.