Fix flaky execution_test
The execution_test was flaky because the "daemon" service exited immediately
after writing to stdout and stderr.  In most cases, the executor would notice
that the pipe is closed first, but in the more seldom case it would discover
that the process has exited and treat it as an error.

For now, let's restore the previous behaviour, where we just ignore it if the
process exits early.

PiperOrigin-RevId: 606212811