Make it possible to signal "impossible" conditions
This is a lightweight way to experimentally probe for dead code, or code which
should be very seldomly run.  For example, one might add:

    should.Never("timeout during operation XYZ")

and then after deployment we should be able to monitor how often this happened.

The open source version has an empty implementation.  The Google-internal
version increments a counter which lets us observe that.  The right variant is
picked at compile time through build constraints.

PiperOrigin-RevId: 606247072