TestStatsMsg: Increase test timeout to de-flake OSX test runs
Getting process resource usage (GetResourceUsageForPID()) can sometimes take
longer (I've measured around ~100ms for that function sometimes), and that does
not play nicely with the timer reset in the resource usage monitor.  In
production deployments, timeouts are much larger than in these tests, so this is
not a problem there.

PiperOrigin-RevId: 607296408