Clean up process lifecycle management in the `entry` package
Before, the interrupt signalling logic (in the `wait_*.go` files) would stop the client and/or process directly.

We move this logic to the `RunMain` functions instead, and create a context that gets canceled under the same conditions.

This context is then passed to the `innerMain` function, which creates a client just like before, but binds the client's lifetime to the context.

This way, we have a more top-down approach, and remove the entry package's dependency on the client.

This allows an easy implementation of config reloading in subsequent contributions.

PiperOrigin-RevId: 667568209