Fix asymmetric path escaping for file server request URLs
The path components are already escaped twice on the client side before making the request. Once explicitly using `url.PathEscape()` on each component, and then by using `url.String()` which escapes the entire path again.

On the server side, using `url.EscapedPath()` instead of `url.Path` caused the latter escaping step on the client side not to be unescaped.

If a file in the filestore contains a special character (e.g. `/`), it could not be accessed due to this bug.

PiperOrigin-RevId: 811375109